+++ src/modules/Modules.test.js
@@ -1,13 +1,14 @@
 import { describe, it, expect } from 'vitest';
 import SomeController from '../controllers/test/SomeController.ts';
 import AbstractController from './AbstractController.ts';
+import { appInstance } from '../helpers/appInstance.ts';
 
 describe('abstract controller methods', () => {
   it('can get routes', async () => {
     expect.assertions(2);
 
-    const controller = new AbstractController(global.server.app);
-    const childController = new SomeController(global.server.app);
+    const controller = new AbstractController(appInstance);
+    const childController = new SomeController(appInstance);
 
     const { routes } = controller;
     const { routes: childRoutes } = childController;